rate limiting, audit logging, 12h times, cleaner AI output, richer patient context
- Rate limit chat endpoints (30/min text, 10/min image)
- Audit log all AI actions (messages, feedback, delete, doctor view)
- Fix audit_log.id to INTEGER for SQLite autoincrement
- Show meal timestamps so AI can identify the most recent meal
- Convert all datetimes to 12-hour format (%I:%M %p)
- Strip asterisks/emoji/━━━ from system prompt for clean AI output
- Include patient name, age, gender, diabetes type in AI context
- Rewrite _build_patient_context with glucose classification, eHbA1c, BMI, variability
- Fix offset-naive vs offset-aware datetime crash in screening section

+++ app/models/audit_log.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .database import Base
@@ -9,7 +9,7 @@ from .database import Base
 class AuditLog(Base):
     __tablename__ = "audit_log"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int | None] = mapped_column(
         Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True
     )
